requirements.txt
websocket-client
asyncio